+++ packages/entities/entities-plugins/src/components/PluginForm.cy.ts
@@ -17,7 +17,7 @@ import schemaMocking from '../../fixtures/schemas/mocking'
 import schemaRateLimiting from '../../fixtures/schemas/rate-limiting'
 import PluginForm from './PluginForm.vue'
 import { PLUGIN_METADATA } from '../definitions/metadata'
-import { EXPERIMENTAL_FREE_FORM_PROVIDER } from '../constants'
+import { EXPERIMENTAL_FREE_FORM_PROVIDER, FEATURE_FLAGS } from '../constants'
 
 const baseConfigKonnect: KonnectPluginFormConfig = {
   app: 'konnect',
@@ -984,6 +984,43 @@ describe('<PluginForm />', () => {
 
         cy.wait('@updatePlugin').then(() => {
           cy.get('@onUpdateSpy').should('have.been.calledOnce')
+        })
+      })
+    })
+
+    describe('Condition field', () => {
+      ;[
+        {
+          description: 'should not show condition field when feature flag is disabled',
+          flagValue: false,
+          assertion: 'not.exist',
+        },
+        {
+          description: 'should show condition field when feature flag is enabled',
+          flagValue: true,
+          assertion: 'exist',
+        },
+      ].forEach(({ description, flagValue, assertion }) => {
+        it(description, () => {
+          interceptKMSchema()
+          const pluginType = 'cors'
+
+          cy.mount(PluginForm, {
+            props: {
+              config: baseConfigKM,
+              pluginType,
+            },
+            global: {
+              provide: {
+                [FEATURE_FLAGS.KM_2306_CONDITION_FIELD_314]: flagValue,
+              },
+            },
+            router,
+          })
+
+          cy.wait('@getPluginSchema')
+          cy.get('.kong-ui-entities-plugin-form-container').should('be.visible')
+          cy.get('#condition').should(assertion)
         })
       })
     })
@@ -2025,6 +2062,43 @@ describe('<PluginForm />', () => {
 
       // VFG should be used for non-freeform plugins
       cy.get('.vue-form-generator').should('exist')
+    })
+
+    describe('Condition field', () => {
+      ;[
+        {
+          description: 'should not show condition field when feature flag is disabled',
+          flagValue: false,
+          assertion: 'not.exist',
+        },
+        {
+          description: 'should show condition field when feature flag is enabled',
+          flagValue: true,
+          assertion: 'exist',
+        },
+      ].forEach(({ description, flagValue, assertion }) => {
+        it(description, () => {
+          interceptKonnectSchema()
+          const pluginType = 'cors'
+
+          cy.mount(PluginForm, {
+            props: {
+              config: baseConfigKonnect,
+              pluginType,
+            },
+            global: {
+              provide: {
+                [FEATURE_FLAGS.KM_2306_CONDITION_FIELD_314]: flagValue,
+              },
+            },
+            router,
+          })
+
+          cy.wait('@getPluginSchema')
+          cy.get('.kong-ui-entities-plugin-form-container').should('be.visible')
+          cy.get('#condition').should(assertion)
+        })
+      })
     })
   })
 })