+++ packages/entities/entities-plugins/src/components/PluginForm.vue
@@ -197,6 +197,7 @@ import unset from 'lodash-es/unset'
 import { REDIS_PARTIAL_INFO } from '../components/free-form/shared/const'
 import type { GlobalAction } from './free-form/shared/types'
 import { FEATURE_FLAGS } from '@kong-ui-public/entities-shared'
+import { FEATURE_FLAGS as PLUGIN_FEATURE_FLAGS } from '../constants'
 
 type ScopedEntitiesType = 'consumer' | 'route' | 'service' | 'consumer_group'
 type Permissions = 'canRetrieve' | 'canEdit' | 'canDelete'
@@ -204,6 +205,7 @@ type ScopedEntityPermission = Partial<Record<Permissions, boolean>>
 type ScopedEntitiesPermissions = Partial<Record<ScopedEntitiesType, ScopedEntityPermission>>
 
 const enabledNewPluginLayout = inject(FEATURE_FLAGS.KM_1948_PLUGIN_FORM_LAYOUT, computed(() => false))
+const enableConditionField = inject<boolean>(PLUGIN_FEATURE_FLAGS.KM_2306_CONDITION_FIELD_314, false)
 
 const emit = defineEmits<{
   (e: 'cancel'): void
@@ -347,6 +349,7 @@ const { customSchemas, typedefs } = composables.useSchemas({
   experimentalRenders: props.config.app === 'konnect' ? props.config.experimentalRenders : undefined,
 })
 const { formatPluginFieldLabel } = composables.usePluginHelpers()
+const { shouldUseFreeForm: isFreeForm } = composables.useFreeFormResolver()
 const { getMessageFromError } = useErrors()
 const { capitalize } = useStringHelpers()
 const { objectsAreEqual } = useHelpers()
@@ -583,6 +586,18 @@ const defaultFormSchema: DefaultPluginsSchemaRecord = reactive({
     help: t('plugins.form.fields.instance_name.help'),
   },
   tags: typedefs.tags as DefaultPluginsFormSchema,
+  ...(
+    enableConditionField ? {
+      condition: {
+        default: '',
+        type: 'input',
+        label: t('plugins.form.fields.condition.label'),
+        inputType: 'text',
+        help: t('plugins.form.fields.condition.help'),
+        placeholder: t('plugins.form.fields.condition.placeholder'),
+      },
+    } : {}
+  ),
 })
 
 // This is specifically used for credential plugins and portal developer plugins
@@ -634,7 +649,7 @@ const buildFormSchema = (parentKey: string, response: Record<string, any>, initi
     // If the field type is 'set', convert it to 'array'
     // Freeform can handle 'set' type with one_of elements as multiselect
     // Todo: create suitable component for 'set' type in freeform and remove this conversion
-    if (scheme.type === 'set' && !(props.engine === 'freeform' && scheme.elements.one_of)) {
+    if (scheme.type === 'set' && !(isFreeForm(props.pluginType, props.engine) && scheme.elements.one_of)) {
       scheme.type = 'array'
     }
     const field = parentKey ? `${parentKey}-${key}` : `${key}`