+++ packages/entities/entities-vaults/sandbox/pages/VaultFormPage.vue
@@ -38,6 +38,7 @@ const konnectConfig = ref<KonnectVaultFormConfig>({
   azureVaultProviderAvailable: true,
   ttl: true,
   hcvAppRoleMethodAvailable: true,
+  hcvCspAuthMethodsAvailable: true,
   hcvSslVerifyAvailable: true,
   conjurVaultProviderAvailable: true,
 })
@@ -55,6 +56,7 @@ const kongManagerConfig = ref<KongManagerVaultFormConfig>({
   azureVaultProviderAvailable: false,
   ttl: true,
   hcvAppRoleMethodAvailable: true,
+  hcvCspAuthMethodsAvailable: true,
   awsStsEndpointUrlAvailable: true,
   conjurVaultProviderAvailable: true,
   base64FieldAvailable: true,